requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.27.2
